+++ examples/resources/dockhand_environment/resource.tf
@@ -6,6 +6,7 @@ resource "dockhand_environment" "example" {
   protocol        = "http"
   tls_skip_verify = false
   icon            = "globe"
+  public_ip       = "203.0.113.10"
 
   collect_activity                 = true
   collect_metrics                  = true